# src/main/java/com/instantsystem/parkingaroundme/api/ParkingsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.instantsystem.parkingaroundme.api;

import com.instantsystem.parkingaroundme.model.ApiErrorResponse;
import com.instantsystem.parkingaroundme.model.ParkingListResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-09-10T15:10:34.653585300+02:00[Europe/Paris]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Parkings", description = "the Parkings API")
public interface ParkingsApi {

    /**
     * GET /api/v1/parkings : Get the list of parking near/around a position
     * give the list of parkings by city and location
     *
     * @param city city identifier (ex poitiers, paris, etc.) (required)
     * @param latitude Lat of the position (optional)
     * @param longitude Long of he position of the user (optional)
     * @param radiusInMeters Search radius in meters around the position (optional, default to 5000)
     * @return List of successfully return parkings (status code 200)
     *         or Invalide request parameters (status code 400)
     *         or Not Found (city not supported or resource not found) (status code 404)
     *         or Internal error (status code 500)
     */
    @Operation(
        operationId = "getNearbyParkings",
        summary = "Get the list of parking near/around a position",
        description = "give the list of parkings by city and location",
        tags = { "Parkings" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of successfully return parkings", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ParkingListResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalide request parameters", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found (city not supported or resource not found)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/parkings",
        produces = { "application/json" }
    )
    
    ResponseEntity<ParkingListResponse> getNearbyParkings(
        @NotNull @Parameter(name = "city", description = "city identifier (ex poitiers, paris, etc.)", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "city", required = true) String city,
        @Parameter(name = "latitude", description = "Lat of the position", in = ParameterIn.QUERY) @Valid @RequestParam(value = "latitude", required = false) Double latitude,
        @Parameter(name = "longitude", description = "Long of he position of the user", in = ParameterIn.QUERY) @Valid @RequestParam(value = "longitude", required = false) Double longitude,
        @Parameter(name = "radiusInMeters", description = "Search radius in meters around the position", in = ParameterIn.QUERY) @Valid @RequestParam(value = "radiusInMeters", required = false, defaultValue = "5000") Integer radiusInMeters
    );

}